refactor(setup): decompose setup-handlers into focused handler modules
Split the 603-line setup-handlers.ts into 13 focused files under
src/setup/handlers/, each well under the 200-line limit. Shared Linear
GraphQL helpers and the SetupApiDeps interface live in shared.ts. The
original setup-handlers.ts is now a thin re-export barrel preserving
backward compatibility for all existing importers.
