+++ src/setup/handlers/index.ts
@@ -1,0 +1,13 @@
+export type { SetupApiDeps } from "./shared.js";
+
+export { handleGetStatus } from "./status.js";
+export { handlePostMasterKey } from "./master-key.js";
+export { handleGetLinearProjects, handlePostLinearProject } from "./linear-project.js";
+export { handlePostOpenaiKey } from "./openai-key.js";
+export { handlePostCodexAuth } from "./codex-auth.js";
+export { handlePostPkceAuthStart, handleGetPkceAuthStatus, handlePostPkceAuthCancel } from "./pkce-auth.js";
+export { handlePostGithubToken } from "./github-token.js";
+export { handlePostCreateTestIssue } from "./test-issue.js";
+export { handlePostCreateLabel } from "./label.js";
+export { handlePostCreateProject } from "./project.js";
+export { handlePostReset } from "./reset.js";
